Converted data storage to binary.

+++ RockSatC-2016/Event Data/BnoData.cs
@@ -5,13 +5,13 @@ namespace RockSatC_2016.Event_Data {
     public class BNOData : IEventData
     {
         public bool loggable => true;
-        public float accel_x { get; set; }
-        public float accel_y { get; set; }
-        public float accel_z { get; set; }
-        public float gyro_x { get; set; }
-        public float gyro_y { get; set; }
-        public float gyro_z { get; set; }
-        public float temp { get; set; }
+        public short accel_x { get; set; }
+        public short accel_y { get; set; }
+        public short accel_z { get; set; }
+        public short gyro_x { get; set; }
+        public short gyro_y { get; set; }
+        public short gyro_z { get; set; }
+        public short temp { get; set; }
 
     }
 }
